Basic background for the hallway
player starts at the leftmost door coming from the mainframe. left is the mag/sheild room. One of the other doors leads to the mystery system (currently unavailible) and the other leads to the UV system (doesn't matter which). to the right is the hover room. up is the sheild room.